*Moved obsolete to partial classes under obsolete-v8.cs *Changed obsolete implementation to throw error as we are not cherry picking these changes any more for 2.3 - will manually add them in instead.
