+++ src/JsonMessageSerializerTest.TestInterfaces.txt
@@ -1,1 +1,1 @@
-﻿{"Data":"dySYxQZnk0oYPM8tKx93VFT9ply/xPk=","S":"kalle","I":42,"B":{"BString":"BOO","C":{"Cstr":"COO"}}}
+﻿{"Data":"FYvPvS75nJ3TZng2dCNicC4Cuds/bec=","S":"kalle","I":42,"B":{"BString":"BOO","C":{"Cstr":"COO"}}}
